--- docs/examples/tables/delete-row.md
@@ -1,15 +1,0 @@
-import { Client, Tables } from "appwrite";
-
-const client = new Client()
-    .setEndpoint('https://<REGION>.cloud.appwrite.io/v1') // Your API Endpoint
-    .setProject('<YOUR_PROJECT_ID>'); // Your project ID
-
-const tables = new Tables(client);
-
-const result = await tables.deleteRow(
-    '<DATABASE_ID>', // databaseId
-    '<TABLE_ID>', // tableId
-    '<ROW_ID>' // rowId
-);
-
-console.log(result);
