+++ src/client.ts
@@ -316,8 +316,8 @@ class Client {
         'x-sdk-name': 'Web',
         'x-sdk-platform': 'client',
         'x-sdk-language': 'web',
-        'x-sdk-version': '19.0.0',
-        'X-Appwrite-Response-Format': '1.8.0',
+        'x-sdk-version': '18.2.0',
+        'X-Appwrite-Response-Format': '1.7.0',
     };
 
     /**
