+++ src/index.ts
@@ -1,7 +1,7 @@
 /**
  * Appwrite Web SDK
  *
- * This SDK is compatible with Appwrite server version 1.8.x. 
+ * This SDK is compatible with Appwrite server version 1.7.x. 
  * For older versions, please check
  * [previous releases](https://github.com/appwrite/sdk-for-web/releases).
  */
@@ -9,7 +9,6 @@ export { Client, Query, AppwriteException } from './client';
 export { Account } from './services/account';
 export { Avatars } from './services/avatars';
 export { Databases } from './services/databases';
-export { Tables } from './services/tables';
 export { Functions } from './services/functions';
 export { Graphql } from './services/graphql';
 export { Locale } from './services/locale';
