+++ src/models.ts
@@ -6,25 +6,11 @@ export namespace Models {
     declare const __default: unique symbol;
 
     /**
-     * Rows List
-     */
-    export type RowList<Row extends Models.Row = Models.DefaultRow> = {
-        /**
-         * Total number of rows rows that matched your query.
-         */
-        total: number;
-        /**
-         * List of rows.
-         */
-        rows: Row[];
-    }
-
-    /**
      * Documents List
      */
     export type DocumentList<Document extends Models.Document = Models.DefaultDocument> = {
         /**
-         * Total number of documents rows that matched your query.
+         * Total number of documents documents that matched your query.
          */
         total: number;
         /**
@@ -38,7 +24,7 @@ export namespace Models {
      */
     export type SessionList = {
         /**
-         * Total number of sessions rows that matched your query.
+         * Total number of sessions documents that matched your query.
          */
         total: number;
         /**
@@ -52,7 +38,7 @@ export namespace Models {
      */
     export type IdentityList = {
         /**
-         * Total number of identities rows that matched your query.
+         * Total number of identities documents that matched your query.
          */
         total: number;
         /**
@@ -66,7 +52,7 @@ export namespace Models {
      */
     export type LogList = {
         /**
-         * Total number of logs rows that matched your query.
+         * Total number of logs documents that matched your query.
          */
         total: number;
         /**
@@ -80,7 +66,7 @@ export namespace Models {
      */
     export type FileList = {
         /**
-         * Total number of files rows that matched your query.
+         * Total number of files documents that matched your query.
          */
         total: number;
         /**
@@ -94,7 +80,7 @@ export namespace Models {
      */
     export type TeamList<Preferences extends Models.Preferences = Models.DefaultPreferences> = {
         /**
-         * Total number of teams rows that matched your query.
+         * Total number of teams documents that matched your query.
          */
         total: number;
         /**
@@ -108,7 +94,7 @@ export namespace Models {
      */
     export type MembershipList = {
         /**
-         * Total number of memberships rows that matched your query.
+         * Total number of memberships documents that matched your query.
          */
         total: number;
         /**
@@ -122,7 +108,7 @@ export namespace Models {
      */
     export type ExecutionList = {
         /**
-         * Total number of executions rows that matched your query.
+         * Total number of executions documents that matched your query.
          */
         total: number;
         /**
@@ -136,7 +122,7 @@ export namespace Models {
      */
     export type CountryList = {
         /**
-         * Total number of countries rows that matched your query.
+         * Total number of countries documents that matched your query.
          */
         total: number;
         /**
@@ -150,7 +136,7 @@ export namespace Models {
      */
     export type ContinentList = {
         /**
-         * Total number of continents rows that matched your query.
+         * Total number of continents documents that matched your query.
          */
         total: number;
         /**
@@ -164,7 +150,7 @@ export namespace Models {
      */
     export type LanguageList = {
         /**
-         * Total number of languages rows that matched your query.
+         * Total number of languages documents that matched your query.
          */
         total: number;
         /**
@@ -178,7 +164,7 @@ export namespace Models {
      */
     export type CurrencyList = {
         /**
-         * Total number of currencies rows that matched your query.
+         * Total number of currencies documents that matched your query.
          */
         total: number;
         /**
@@ -192,7 +178,7 @@ export namespace Models {
      */
     export type PhoneList = {
         /**
-         * Total number of phones rows that matched your query.
+         * Total number of phones documents that matched your query.
          */
         total: number;
         /**
@@ -206,7 +192,7 @@ export namespace Models {
      */
     export type LocaleCodeList = {
         /**
-         * Total number of localeCodes rows that matched your query.
+         * Total number of localeCodes documents that matched your query.
          */
         total: number;
         /**
@@ -214,51 +200,6 @@ export namespace Models {
          */
         localeCodes: LocaleCode[];
     }
-
-    /**
-     * Row
-     */
-    export type Row = {
-        /**
-         * Row ID.
-         */
-        $id: string;
-        /**
-         * Row automatically incrementing ID.
-         */
-        $sequence: number;
-        /**
-         * Table ID.
-         */
-        $tableId: string;
-        /**
-         * Database ID.
-         */
-        $databaseId: string;
-        /**
-         * Row creation date in ISO 8601 format.
-         */
-        $createdAt: string;
-        /**
-         * Row update date in ISO 8601 format.
-         */
-        $updatedAt: string;
-        /**
-         * Row permissions. [Learn more about permissions](https://appwrite.io/docs/permissions).
-         */
-        $permissions: string[];
-    }
-
-    export type DefaultRow = Row & {
-        [key: string]: any;
-        [__default]: true;
-    };
-
-    export type DataWithoutRowKeys = {
-        [K in string]: any;
-    } & {
-        [K in keyof Row]?: never;
-    };
 
     /**
      * Document
