+++ src/services/databases.ts
@@ -17,7 +17,6 @@ export class Databases {
      * @param {string[]} queries
      * @throws {AppwriteException}
      * @returns {Promise<Models.DocumentList<Document>>}
-     * @deprecated This API has been deprecated since 1.8.0. Please use `Tables.listRows` instead.
      */
     listDocuments<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, queries?: string[]): Promise<Models.DocumentList<Document>> {
         if (typeof databaseId === 'undefined') {
@@ -54,7 +53,6 @@ export class Databases {
      * @param {string[]} permissions
      * @throws {AppwriteException}
      * @returns {Promise<Document>}
-     * @deprecated This API has been deprecated since 1.8.0. Please use `Tables.createRow` instead.
      */
     createDocument<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string, data: Document extends Models.DefaultDocument ? Models.DataWithoutDocumentKeys : Omit<Document, keyof Models.Document>, permissions?: string[]): Promise<Document> {
         if (typeof databaseId === 'undefined') {
@@ -103,7 +101,6 @@ export class Databases {
      * @param {string[]} queries
      * @throws {AppwriteException}
      * @returns {Promise<Document>}
-     * @deprecated This API has been deprecated since 1.8.0. Please use `Tables.getRow` instead.
      */
     getDocument<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string, queries?: string[]): Promise<Document> {
         if (typeof databaseId === 'undefined') {
@@ -141,11 +138,12 @@ export class Databases {
      * @param {string} databaseId
      * @param {string} collectionId
      * @param {string} documentId
+     * @param {object} data
+     * @param {string[]} permissions
      * @throws {AppwriteException}
      * @returns {Promise<Document>}
-     * @deprecated This API has been deprecated since 1.8.0. Please use `Tables.upsertRow` instead.
      */
-    upsertDocument<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string): Promise<Document> {
+    upsertDocument<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string, data: object, permissions?: string[]): Promise<Document> {
         if (typeof databaseId === 'undefined') {
             throw new AppwriteException('Missing required parameter: "databaseId"');
         }
@@ -155,8 +153,17 @@ export class Databases {
         if (typeof documentId === 'undefined') {
             throw new AppwriteException('Missing required parameter: "documentId"');
         }
+        if (typeof data === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "data"');
+        }
         const apiPath = '/databases/{databaseId}/collections/{collectionId}/documents/{documentId}'.replace('{databaseId}', databaseId).replace('{collectionId}', collectionId).replace('{documentId}', documentId);
         const payload: Payload = {};
+        if (typeof data !== 'undefined') {
+            payload['data'] = data;
+        }
+        if (typeof permissions !== 'undefined') {
+            payload['permissions'] = permissions;
+        }
         const uri = new URL(this.client.config.endpoint + apiPath);
 
         const apiHeaders: { [header: string]: string } = {
@@ -181,7 +188,6 @@ export class Databases {
      * @param {string[]} permissions
      * @throws {AppwriteException}
      * @returns {Promise<Document>}
-     * @deprecated This API has been deprecated since 1.8.0. Please use `Tables.updateRow` instead.
      */
     updateDocument<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string, data?: Partial<Document extends Models.DefaultDocument ? Models.DataWithoutDocumentKeys : Omit<Document, keyof Models.Document>>, permissions?: string[]): Promise<Document> {
         if (typeof databaseId === 'undefined') {
@@ -223,7 +229,6 @@ export class Databases {
      * @param {string} documentId
      * @throws {AppwriteException}
      * @returns {Promise<{}>}
-     * @deprecated This API has been deprecated since 1.8.0. Please use `Tables.deleteRow` instead.
      */
     deleteDocument(databaseId: string, collectionId: string, documentId: string): Promise<{}> {
         if (typeof databaseId === 'undefined') {
@@ -245,6 +250,100 @@ export class Databases {
 
         return this.client.call(
             'delete',
+            uri,
+            apiHeaders,
+            payload
+        );
+    }
+
+    /**
+     * Decrement a specific attribute of a document by a given value.
+     *
+     * @param {string} databaseId
+     * @param {string} collectionId
+     * @param {string} documentId
+     * @param {string} attribute
+     * @param {number} value
+     * @param {number} min
+     * @throws {AppwriteException}
+     * @returns {Promise<Document>}
+     */
+    decrementDocumentAttribute<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string, attribute: string, value?: number, min?: number): Promise<Document> {
+        if (typeof databaseId === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "databaseId"');
+        }
+        if (typeof collectionId === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "collectionId"');
+        }
+        if (typeof documentId === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "documentId"');
+        }
+        if (typeof attribute === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "attribute"');
+        }
+        const apiPath = '/databases/{databaseId}/collections/{collectionId}/documents/{documentId}/{attribute}/decrement'.replace('{databaseId}', databaseId).replace('{collectionId}', collectionId).replace('{documentId}', documentId).replace('{attribute}', attribute);
+        const payload: Payload = {};
+        if (typeof value !== 'undefined') {
+            payload['value'] = value;
+        }
+        if (typeof min !== 'undefined') {
+            payload['min'] = min;
+        }
+        const uri = new URL(this.client.config.endpoint + apiPath);
+
+        const apiHeaders: { [header: string]: string } = {
+            'content-type': 'application/json',
+        }
+
+        return this.client.call(
+            'patch',
+            uri,
+            apiHeaders,
+            payload
+        );
+    }
+
+    /**
+     * Increment a specific attribute of a document by a given value.
+     *
+     * @param {string} databaseId
+     * @param {string} collectionId
+     * @param {string} documentId
+     * @param {string} attribute
+     * @param {number} value
+     * @param {number} max
+     * @throws {AppwriteException}
+     * @returns {Promise<Document>}
+     */
+    incrementDocumentAttribute<Document extends Models.Document = Models.DefaultDocument>(databaseId: string, collectionId: string, documentId: string, attribute: string, value?: number, max?: number): Promise<Document> {
+        if (typeof databaseId === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "databaseId"');
+        }
+        if (typeof collectionId === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "collectionId"');
+        }
+        if (typeof documentId === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "documentId"');
+        }
+        if (typeof attribute === 'undefined') {
+            throw new AppwriteException('Missing required parameter: "attribute"');
+        }
+        const apiPath = '/databases/{databaseId}/collections/{collectionId}/documents/{documentId}/{attribute}/increment'.replace('{databaseId}', databaseId).replace('{collectionId}', collectionId).replace('{documentId}', documentId).replace('{attribute}', attribute);
+        const payload: Payload = {};
+        if (typeof value !== 'undefined') {
+            payload['value'] = value;
+        }
+        if (typeof max !== 'undefined') {
+            payload['max'] = max;
+        }
+        const uri = new URL(this.client.config.endpoint + apiPath);
+
+        const apiHeaders: { [header: string]: string } = {
+            'content-type': 'application/json',
+        }
+
+        return this.client.call(
+            'patch',
             uri,
             apiHeaders,
             payload
